pyproject.toml
[build-system]
requires = ["setuptools>=69", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_file = "src/aspy21/_version.py"

[project]
name = "aspy21"
dynamic = ["version"]
description = "Modern Python client for Aspen InfoPlus.21 (IP.21) with pandas DataFrame output, tag search, and automatic retry logic"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{ name = "bazdalaz" }]
maintainers = [{ name = "bazdalaz", email = "bazdalaz@users.noreply.github.com" }]
keywords = ["aspen", "ip21", "historian", "process data", "industrial", "ot", "scada", "aspentech"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
  "httpx>=0.27",
  "pandas>=2.0",
  "tenacity>=9.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "pytest-asyncio>=0.23",
  "respx>=0.21",           # mock for httpx
  "ruff>=0.8",
  "pyright>=1.1.389",
  "types-python-dateutil",
  "pre-commit>=3.5",
  "build>=1.0",            # build distribution packages
  "twine>=5.0",            # upload to PyPI
  "setuptools-scm>=8"      # for bump_version.py script
]
examples = [
  "python-dotenv>=1.0"     # for .env file support in examples
]

[project.urls]
Homepage = "https://github.com/bazdalaz/aspy21"
Repository = "https://github.com/bazdalaz/aspy21"
Issues = "https://github.com/bazdalaz/aspy21/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["aspy21*"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py39"

[tool.ruff.lint]
extend-select = ["I", "B", "UP", "SIM", "C4", "F", "E", "W"]
ignore = ["D100", "D104", "D105"]
exclude = ["examples"]

[tool.ruff.lint.extend-per-file-ignores]
"examples/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=src/aspy21",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v"
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__", ".venv", "build", "dist", "scripts"]
pythonVersion = "3.9"
typeCheckingMode = "standard"
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"
ignore = ["examples"]
